Fixed overflow in temperature by increasing size of read_AT30TSE758 to int32, whilst keeping temperature at uint16 as assue its not negative. Note there is a multiplication by 125 in read_AT30TSE758, hence the need for 32bit to contain the temperature on the range -55.000 to +125.000 (ie -55000 to +125000)

+++ temperature.h
@@ -34,7 +34,7 @@
 /** @brief      Temperate measure, hardware agnostic abstraction
     @retval     temperature Temperate in °Cx1000
 */
-int16_t temperature(void);
+uint16_t temperature(void);
 
 /** @brief      The change in temperature over one second
     @retval     delta_t The change in temperature over one second in °Cx1000
@@ -66,4 +66,4 @@ uint8_t init_AT30TSE758(uint8_t address);
     @retval     temperature     in °Cx1000, or
     @retval     0xFFFF          if read failed
 */
-int16_t read_AT30TSE758(uint8_t address);
+int32_t read_AT30TSE758(uint8_t address);
